pipeline/requirements.txt
pandas
googlemaps
sqlalchemy
psycopg2-binary
dotenv
delta-spark
pyspark
requests
confluent-kafka